% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bssa_2014_nga}
\alias{bssa_2014_nga}
\title{The GMPE of BSSA 2014}
\usage{
bssa_2014_nga(M, period, Rjb, Fault_Type, region, z1, Vs30, coeffs)
}
\arguments{
\item{M}{A list of moment magnitudes for all interest of events}

\item{period}{The period of interest. -1 for PGV, 0 for PGA. A numeric value.}

\item{Rjb}{A list of Joyner-Boore distance (km); closest distance (km) to surface
projection of rupture plane. The length should equal to the length of moment magnitudes}

\item{Fault_Type}{A list of fault type indicators: 0 for unspecified fault;
1 for strike-slip fault; 2 for normal fault; 3 for reverse fault.
The length should equal to the length of moment magnitudes}

\item{region}{A list of region indicators: 0 for global (incl. Taiwan); 1 for California;
2 for Japan; 3 for China or Turkey; 4 for Italy. The length should equal to the length of moment magnitudes}

\item{z1}{Basin depth (km): depth from the ground surface to the 1km/s shear-wave horizon.
-999 if unknown. A numeric value.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). A numeric value.}

\item{coeffs}{The coefficient table of BSSA 2014. You can use the internal saved data object, bssa_2014_coeffs}
}
\value{
A dataframe with four calculated columns, median/sigma/phi/tau. Each row is the calcualted values for each event.
}
\description{
This function calculates the ground motion median values and standard deviations
}
\references{
Boore, D. M., Stewart, J. P., Seyhan, E., and Atkinson, G. M. (2014).
NGA-West2 Equations for Predicting PGA, PGV, and 5\% Damped PSA for
Shallow Crustal Earthquakes. Earthquake Spectra, 30(3), 1057-1085.
}
