% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haz_calc.R
\name{event_haz_calc}
\alias{event_haz_calc}
\title{The subroutine function for hazard curve calculation at a site from a given event}
\usage{
event_haz_calc(Mag, AORate, Fault_type, region, periods, IM_mat, Rjb, Vs30, z1)
}
\arguments{
\item{Mag}{The magnitude of the event}

\item{AORate}{The annual occurrence rate of the event}

\item{Fault_type}{An indicator for fault type : 0 for unspecified fault;
1 for strike-slip fault; 2 for normal fault; 3 for reverse fault.}

\item{region}{An indicator for region: 0 for global (incl. Taiwan); 1 for California;
2 for Japan; 3 for China or Turkey; 4 for Italy.}

\item{periods}{An interest period or period arrays (0 for PGA, -1 for PGV)}

\item{IM_mat}{A vector or a matrix that specifies the IM exceedance levels for
hazard curve calculation.
If \code{periods} length is 1, then \code{IM_mat} is a vector;
otherwise, \code{IM_mat} is a matrix with row number equals to \code{periods} length.
Each row in \code{IM_mat} is the specified IM exceedance levels for each of IM in \code{periods}.}

\item{Rjb}{The Rjb distance from the event to the interest site}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). A numeric value.}

\item{z1}{Basin depth (km): depth from the ground surface to the 1km/s shear-wave horizon.
-999 if unknown. A numeric value.}
}
\value{
A hazard curve matrix at the given IM levels for the event
}
\description{
This is a subroutine function to calculate the hazard curve (annual exceedance rate curve)
at a site from a given event
}
