% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_selection.R
\name{event_selection}
\alias{event_selection}
\title{A function for event selection}
\usage{
event_selection(Y, X, target_num = NULL, max_amp = NULL, min_hazard = 1e-07)
}
\arguments{
\item{Y}{The target hazard and target hazard magnitude/distance distributions}

\item{X}{The hazard matrix produced by each of the considered events}

\item{target_num}{The number of final selected event subset, a positive integer}

\item{max_amp}{The maximum amplification adjusted factor, a positive numeric number. It is only
working when target_num is not set}

\item{min_hazard}{The minimum hazard value in \code{Y} that is considered in regression.
This is used to avoid having extremely large weights. Note the weights is 1/\code{Y}.
The default value for \code{min_hazard} is 1e-7.}
}
\value{
A list of three elements is returned. The first element is
the corresponding adjusted amplification factors to each event
(or each column in \code{X}), the second element is recovered hazard by the selected subset events,
and the last element is the logarithmic mean squared error.
The events with positive factors are selected events and the events with factors = 0 are
NOT selected.
}
\description{
This function takes the target hazard (and target hazard magnitude and/or distance distributions)
and the hazard (and hazard magnitude and/or distance distributions) produced by each of the
considered events to efficiently and optimally select a subset events with adjusted annual
occurrence rates by using LASSO regression.
}
