% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haz_calc.R
\name{events_hazmat_calc}
\alias{events_hazmat_calc}
\title{The function to construct the hazard matrix given event set and site list}
\usage{
events_hazmat_calc(Y_haz, EventSet, SiteTable)
}
\arguments{
\item{Y_haz}{The target hazard. The column names and data format should be consistent with
the example on this page:
\url{https://raw.githubusercontent.com/wltcwpf/Dataset/main/RPSHA/Y_Haz.csv}}

\item{EventSet}{The event set. The column names and data format should be consistent with
the example on this page:
\url{https://raw.githubusercontent.com/wltcwpf/Dataset/main/RPSHA/EventSet.csv}}

\item{SiteTable}{The site list file set. The column names and data format should be consistent with
the example on this page:
\url{https://raw.githubusercontent.com/wltcwpf/Dataset/main/RPSHA/SiteTable.csv}}
}
\value{
A dataframe with hazard curves produced by each event in \code{EventSet} is returned.
An example result can be found on this shared Google file:
\url{https://drive.google.com/file/d/1b_vxrXVSW4h0u_hzcrLif6CvQthj2ZO1/view?usp=sharing}
}
\description{
This is a function to calculate the hazard curves (annual exceedance rate curve)
at all sites from a event set
}
