% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM32_Secs.R
\docType{data}
\name{FM32_Secs}
\alias{FM32_Secs}
\title{The UCERF3 3.2 branch average model for parent fault sub-sections}
\format{
A data frame with 2664 rows (number of all modeled fault sub-sections) and 8 columns
(2 columns for the unique index and sub-section name; following by area, depth, dip, dip direction,
subsection bottom depth, and the traces).
}
\source{
USGS github repository: \url{https://github.com/usgs/nshm-cous-2014}

UCERF3 Reference: \url{https://doi.org/10.1785/0120130164}
}
\usage{
FM32_Secs
}
\description{
A dataframe containing 3.2 branch average model for all parent fault sub-sections from UCERF3.
The data is originally from USGS nshm-cous-2014 github repository.
}
\keyword{datasets}
