% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haz_calc.R
\name{Haz_calc}
\alias{Haz_calc}
\title{The hazard calculator}
\usage{
Haz_calc(
  slat,
  slon,
  Vs30,
  z1 = -999,
  max_dist = 300,
  periods,
  IM_mat = NA,
  site_name = "Site1",
  output_dir
)
}
\arguments{
\item{slat}{The latitude of the interest site}

\item{slon}{The longitude of the interest site}

\item{Vs30}{The associated Vs30 value (in m/s) to the interest site}

\item{z1}{The associated z1 value (in km) to the interest site. -999 for unknown.}

\item{max_dist}{The maximum cut-off distance for considered sources. Default is 300 km.}

\item{periods}{The list of interest IMs. Use -1 for PGV, 0 for PGA, and other NGA-West2 periods for PSA}

\item{IM_mat}{A matrix that specifies the IM exceedance levels for hazard curve calculation.
If \code{periods} length is 1, then \code{IM_mat} is a vector;
otherwise, \code{IM_mat} is a matrix with row number equals to \code{periods} length.
Each row in \code{IM_mat} is the specified IM exceedance levels for each of IM in \code{periods}.
The default is NA, which uses the default IM exceedance levels.}

\item{site_name}{The name of the site}

\item{output_dir}{The directory for the outputs}
}
\description{
This function calculates hazard for the given IMs at the user-defined return periods or IM levels at
the given sites using UCERF3 3.1 and 3.2 branches average source model.
}
\examples{
Haz_calc(slat = 36.1, slon = -118.1, Vs30 = 350,
z1 = -999, periods = c(-1, 0.01, 5), output_dir = '~/Desktop/RPSHA')

Haz_calc(slat = 36.5, slon = -119.5, Vs30 = 567,
z1 = 1.5, periods = c(0.01, 1, 5),
IM_mat = rbind(c(0.001, 0.0016), c(0.001, 0.0016), c(0.001, 0.0016)), site_name = 'Site2',
output_dir = '~/Desktop/RPSHA')

}
