% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haz_calc.R
\name{Haz_calc_sub}
\alias{Haz_calc_sub}
\title{The subroutine of hazard curve calculation}
\usage{
Haz_calc_sub(df_source, source_type, period, IM, z1, Vs30, weight = 1)
}
\arguments{
\item{df_source}{The returned dataframe by \code{flts_filter} and \code{pts_filter} functions}

\item{source_type}{The indicator of source type: 0 for flts; 1 for pts}

\item{period}{The interest period}

\item{IM}{IM level array}

\item{z1}{Basin depth (km): depth from the ground surface to the 1km/s shear-wave horizon.
-999 if unknown. A numeric value.}

\item{Vs30}{Shear wave velocity averaged over top 30 m (in m/s). A numeric value.}

\item{weight}{The weight that is supposed to be considered for the hazard}
}
\value{
A hazard matrix at the given IM levels for each event in \code{df_sources}
}
\description{
This is a subroutine of \code{haz_calc} function
}
